+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/backend/data/DataState.kt
@@ -1,9 +1,0 @@
-package com.github.warningimhack3r.npmupdatedependencies.backend.data
-
-import kotlinx.datetime.Instant
-
-data class DataState<T>(
-    val data: T?,
-    val scannedAt: Instant,
-    val comparator: String
-)