+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/ui/helpers/AnnotatorsCommon.kt
@@ -1,7 +1,7 @@
 package com.github.warningimhack3r.npmupdatedependencies.ui.helpers
 
-import com.github.warningimhack3r.npmupdatedependencies.backend.data.Property
 import com.github.warningimhack3r.npmupdatedependencies.backend.extensions.stringValue
+import com.github.warningimhack3r.npmupdatedependencies.backend.models.Property
 import com.intellij.json.psi.JsonProperty
 import com.intellij.psi.PsiFile
 import com.intellij.psi.util.PsiTreeUtil